fix before subscription

+++ include/feetech_utils.hpp
@@ -18,7 +18,7 @@ void feetechWrite(SMS_STS &st, unsigned char id,
     st.RegWritePosEx(id, pos_cnt, speed_cnt, acc_cnt);
 }
 
-void wakeup_sweep(SMS_STS &st, float angles[], int length_angle, int ids[], int length_id, int interval = 1000)
+void wakeup_sweep(SMS_STS &st, float angles[], size_t length_angle, int ids[], size_t length_id, int interval = 1000)
 {
     for (size_t j = 0; j < length_angle; j++)
     {
